Clean up tests, move setup into its own file
